add useSocialEvents hook for social event handling
